requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
# python -m pip install -r requirements.txt
azure-functions==1.7.2
azure-functions-durable>=1.0.0b12
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
torchvision==0.7.0+cpu
jsonpickle==2.0.0
adaptnlp==0.2.2
flair==0.7
nltk==3.6.2
numpy==1.20.3
openpyxl==3.0.7
#bertopic==0.8.1
bertopic==0.9
sklearn==0.0
psutil==5.8.0
tomotopy==0.12.1
#spacy # Read how to do for Azure at https://williamandrewgriffin.com/best-way-to-deploy-spacy-to-azure-functions/
psycopg2==2.8.6
urllib3==1.26.5
transformers==3.5.1
